Wednesday debug

+++ src/buffer_data.c
@@ -1,8 +1,8 @@
 #include "common.h"
 #include "buffer_data.h"
 
-float audio_bufferR[BUFFER_SIZE] = {0};
-float audio_bufferL[BUFFER_SIZE] = {0};
+short audio_bufferR[BUFFER_SIZE] = {0};
+short audio_bufferL[BUFFER_SIZE] = {0};
 
 // fftw_complex audio_fft_bufferR[BUFFER_SIZE];
 // fftw_complex audio_fft_bufferL[BUFFER_SIZE];
